refactor: useCookieUserInfo hook 적용 방식 변경
